+++ src/app/product/page.tsx
@@ -1,11 +1,9 @@
-"use client";
 import { Button } from "@/components/ui/button";
 
 export default function Home() {
   return (
     <div className="font-sans grid grid-rows-[20px_1fr_20px] items-center justify-items-center min-h-screen p-8 pb-20 gap-16 sm:p-20">
       Product
-      <Button onClick={() => alert("simboraaaa")}>clica aqui</Button>
     </div>
   );
 }
